ndx_options/requirements.txt
ib_insync>=0.9.86
numpy>=1.24
pandas>=2.0
scipy>=1.11
python-dotenv>=1.0
